chat-bot/requirements.txt
fastapi 
uvicorn 
requests 
openai 
qdrant-client
python-dotenv
ragas 
datasets
langchain-openai